Move diagnostic builder tests to LangServer tests and clean up usings
